+++ pioreactor/tests/test_calibrations.py
@@ -124,21 +124,21 @@ def test_predict_linear(calibration) -> None:
     calibration.curve_data_ = [3, 2]  # 3x + 2
     x = 4
     expected_y = 3 * x + 2
-    assert calibration.predict(x) == expected_y
+    assert calibration.x_to_y(x) == expected_y
 
 
 def test_predict_quadratic(calibration) -> None:
     calibration.curve_data_ = [5, 3, 2]  # 5x^2 + 3x + 2
     x = 2
     expected_y = 5 * x**2 + 3 * x + 2
-    assert calibration.predict(x) == expected_y
+    assert calibration.x_to_y(x) == expected_y
 
 
 def test_ipredict_linear(calibration) -> None:
     calibration.curve_data_ = [3, 2]  # 3x + 2
     y = 14
     expected_x = (y - 2) / 3
-    assert calibration.ipredict(y) == pytest.approx(expected_x)
+    assert calibration.y_to_x(y) == pytest.approx(expected_x)
 
 
 def test_ipredict_quadratic_single_solution(calibration) -> None:
@@ -146,20 +146,20 @@ def test_ipredict_quadratic_single_solution(calibration) -> None:
     calibration.recorded_data = {"x": [0, 2], "y": [2, 20]}
     y = 12
     expected_x = 1.145683229480096  # Solves 5x^2 + 3x + 2 = 12
-    assert calibration.ipredict(y) == pytest.approx(expected_x)
+    assert calibration.y_to_x(y) == pytest.approx(expected_x)
 
 
 def test_ipredict_no_solution(calibration) -> None:
     calibration.curve_data_ = [1, 0, 5]  # x^2 + 5, no solution for y = -10
     with pytest.raises(exc.NoSolutionsFoundError):
-        calibration.ipredict(-10)
+        calibration.y_to_x(-10)
 
 
 def test_ipredict_multiple_solutions(calibration) -> None:
     calibration.curve_data_ = [1, 0, -6]  # x^2 - 6, solutions for y=0 are +- 2.45
     calibration.recorded_data = {"x": [0, 3], "y": [0, 9]}
     y = 0
-    assert calibration.ipredict(y) == pytest.approx(2.44948974)
+    assert calibration.y_to_x(y) == pytest.approx(2.44948974)
 
 
 def test_ipredict_solution_below_domain(calibration) -> None:
@@ -167,7 +167,7 @@ def test_ipredict_solution_below_domain(calibration) -> None:
     calibration.recorded_data = {"x": [0, 1], "y": [10, 20]}
     y = 1.99  # Solution below domain
     with pytest.raises(exc.SolutionBelowDomainError):
-        calibration.ipredict(y, enforce_bounds=True)
+        calibration.y_to_x(y, enforce_bounds=True)
 
 
 def test_ipredict_solution_above_domain(calibration) -> None:
@@ -175,15 +175,15 @@ def test_ipredict_solution_above_domain(calibration) -> None:
     calibration.recorded_data = {"x": [0, 1], "y": [0, 100]}
     y = 50  # Solution above domain
     with pytest.raises(exc.SolutionAboveDomainError):
-        calibration.ipredict(y, enforce_bounds=True)
+        calibration.y_to_x(y, enforce_bounds=True)
 
 
 def test_predict_ipredict_consistency(calibration) -> None:
     calibration.curve_data_ = [2, -3, 1]  # 2x^2 - 3x + 1
     calibration.recorded_data = {"x": [0, 3], "y": [1, 16]}
     x = 2
-    y = calibration.predict(x)
-    assert calibration.ipredict(y) == pytest.approx(x)
+    y = calibration.x_to_y(x)
+    assert calibration.y_to_x(y) == pytest.approx(x)
 
 
 def test_linear_data_produces_linear_curve_in_range_even_if_high_degree() -> None:
@@ -225,8 +225,8 @@ def test_mandys_data_for_pathological_poly() -> None:
         pd_channel="2",
     )
 
-    assert abs(mcal.predict(0.002) - curve_callable(0.002)) < 1e-10
-    assert abs(mcal.ipredict(0.002) - 0.002) < 0.1
+    assert abs(mcal.x_to_y(0.002) - curve_callable(0.002)) < 1e-10
+    assert abs(mcal.y_to_x(0.002) - 0.002) < 0.1
 
 
 def test_custom_protocol():